Refactor orienteering functions as class
The separate functions that comprised the orienteering problem solver
have been unified as a single class in orienteering.py. The new class
loosely follows a scikit-learn paradigm in that the problem is
constructed, then solved, and data is made accessible through class
attributes. Additional minor changes include a preliminary TODO list
and removal of unused or unnecessary code.