Add Tabu Search, remove unused orienteering params
Implemented a Tabu Search-like function with the intention of finding
faster solutions to the orienteering problem. The implementation
diverges slightly from the procedure outlined by Liang et al. in
"Meta Heuristics for the Orienteering Problem" but shares the same
fundamental approach. The stochastic nature of the algorithm means
sacrifices consistency for performance. Doing so has allowed for an
efficient means of finding near optimal paths on the entire set of
locations, greatly out performing the MIPS approach in terms of both
speed and final path result.