Address data leakage in RF, allow solver kwargs
Added a holdout validation set to Random Forest model to verify its
ability to generalize to new coordinates. The model is still trained on
bidirectional locations, but the impact on estimates seems negligible.
Altered Orienteering solve function to allow general keyword arguments
to be passed to the solver. This change provides a means for the solver
to return a suboptimal solution in a realistic timespan.
Added nbviewer links to README.